+++ packages/flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterPlugin.kt
@@ -64,8 +64,6 @@ class SentryFlutterPlugin :
       "setContexts" -> setContexts(call.argument("key"), call.argument("value"), result)
       "removeContexts" -> removeContexts(call.argument("key"), result)
       "setUser" -> setUser(call.argument("user"), result)
-      "addBreadcrumb" -> addBreadcrumb(call.argument("breadcrumb"), result)
-      "clearBreadcrumbs" -> clearBreadcrumbs(result)
       "setExtra" -> setExtra(call.argument("key"), call.argument("value"), result)
       "removeExtra" -> removeExtra(call.argument("key"), result)
       "setTag" -> setTag(call.argument("key"), call.argument("value"), result)
@@ -186,24 +184,6 @@ class SentryFlutterPlugin :
     } else {
       Sentry.setUser(null)
     }
-    result.success("")
-  }
-
-  private fun addBreadcrumb(
-    breadcrumb: Map<String, Any?>?,
-    result: Result,
-  ) {
-    if (breadcrumb != null) {
-      val options = ScopesAdapter.getInstance().options
-      val breadcrumbInstance = Breadcrumb.fromMap(breadcrumb, options)
-      Sentry.addBreadcrumb(breadcrumbInstance)
-    }
-    result.success("")
-  }
-
-  private fun clearBreadcrumbs(result: Result) {
-    Sentry.clearBreadcrumbs()
-
     result.success("")
   }