+++ packages/flutter/lib/src/native/java/sentry_native_java.dart
@@ -242,4 +242,59 @@ class SentryNativeJava extends SentryNativeChannel {
     await _envelopeSender?.close();
     return super.close();
   }
+
+  @override
+  void addBreadcrumb(Breadcrumb breadcrumb) =>
+      tryCatchSync('addBreadcrumb', () {
+        using((arena) {
+          final nativeOptions = native.ScopesAdapter.getInstance()?.getOptions()
+            ?..releasedBy(arena);
+          if (nativeOptions == null) return;
+          final jMap = _dartToJMap(breadcrumb.toJson(), arena);
+          final nativeBreadcrumb =
+              native.Breadcrumb.fromMap(jMap, nativeOptions)
+                ?..releasedBy(arena);
+          if (nativeBreadcrumb == null) return;
+          native.Sentry.addBreadcrumb$1(nativeBreadcrumb);
+        });
+      });
+
+  @override
+  void clearBreadcrumbs() => tryCatchSync('clearBreadcrumbs', () {
+        native.Sentry.clearBreadcrumbs();
+      });
+}
+
+JObject? _dartToJObject(Object? value, Arena arena) => switch (value) {
+      null => null,
+      String s => s.toJString()..releasedBy(arena),
+      bool b => b.toJBoolean()..releasedBy(arena),
+      int i => i.toJLong()..releasedBy(arena),
+      double d => d.toJDouble()..releasedBy(arena),
+      List<dynamic> l => _dartToJList(l, arena),
+      Map<String, dynamic> m => _dartToJMap(m, arena),
+      _ => null
+    };
+
+JList<JObject?> _dartToJList(List<dynamic> values, Arena arena) {
+  final jlist = JList.array(JObject.nullableType)..releasedBy(arena);
+
+  for (final value in values) {
+    final jObj = _dartToJObject(value, arena);
+    jlist.add(jObj);
+  }
+
+  return jlist;
+}
+
+JMap<JString, JObject?> _dartToJMap(Map<String, dynamic> json, Arena arena) {
+  final jmap = JMap.hash(JString.type, JObject.nullableType)..releasedBy(arena);
+
+  for (final entry in json.entries) {
+    final key = entry.key.toJString()..releasedBy(arena);
+    final value = _dartToJObject(entry.value, arena);
+    jmap[key] = value;
+  }
+
+  return jmap;
 }