+++ packages/flutter/test/sentry_native_channel_test.dart
@@ -10,8 +10,8 @@ import 'package:mockito/mockito.dart';
 import 'package:sentry/src/platform/mock_platform.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/native/factory.dart';
-import 'package:sentry_flutter/src/native/method_channel_helper.dart';
 import 'package:sentry_flutter/src/native/sentry_native_binding.dart';
+import 'package:sentry_flutter/src/native/utils/data_normalizer.dart';
 import 'package:sentry_flutter/src/replay/replay_config.dart';
 
 import 'mocks.dart';
@@ -59,7 +59,7 @@ void main() {
           username: user.username,
           email: user.email,
           ipAddress: user.ipAddress,
-          data: MethodChannelHelper.normalizeMap(user.data),
+          data: normalizeMap(user.data),
           // ignore: deprecated_member_use
           extras: user.extras,
           geo: user.geo,
@@ -77,41 +77,36 @@ void main() {
       });
 
       test('addBreadcrumb', () async {
+        final matcher = _nativeUnavailableMatcher(
+          mockPlatform,
+          includeLookupSymbol: true,
+          includeFailedToLoadClassException: true,
+        );
+
         final breadcrumb = Breadcrumb(
           data: {'object': Object()},
         );
-        final normalizedBreadcrumb = Breadcrumb(
-          message: breadcrumb.message,
-          category: breadcrumb.category,
-          data: MethodChannelHelper.normalizeMap(breadcrumb.data),
-          level: breadcrumb.level,
-          type: breadcrumb.type,
-          timestamp: breadcrumb.timestamp,
-          // ignore: invalid_use_of_internal_member
-          unknown: breadcrumb.unknown,
-        );
-        when(channel.invokeMethod(
-                'addBreadcrumb', {'breadcrumb': normalizedBreadcrumb.toJson()}))
-            .thenAnswer((_) => Future.value());
 
-        await sut.addBreadcrumb(breadcrumb);
+        expect(() => sut.addBreadcrumb(breadcrumb), matcher);
 
-        verify(channel.invokeMethod(
-            'addBreadcrumb', {'breadcrumb': normalizedBreadcrumb.toJson()}));
+        verifyZeroInteractions(channel);
       });
 
       test('clearBreadcrumbs', () async {
-        when(channel.invokeMethod('clearBreadcrumbs'))
-            .thenAnswer((_) => Future.value());
+        final matcher = _nativeUnavailableMatcher(
+          mockPlatform,
+          includeLookupSymbol: true,
+          includeFailedToLoadClassException: true,
+        );
 
-        await sut.clearBreadcrumbs();
+        expect(() => sut.clearBreadcrumbs(), matcher);
 
-        verify(channel.invokeMethod('clearBreadcrumbs'));
+        verifyZeroInteractions(channel);
       });
 
       test('setContexts', () async {
         final value = {'object': Object()};
-        final normalizedValue = MethodChannelHelper.normalize(value);
+        final normalizedValue = normalize(value);
         when(channel.invokeMethod('setContexts', {
           'key': 'fixture-key',
           'value': normalizedValue
@@ -134,7 +129,7 @@ void main() {
 
       test('setExtra', () async {
         final value = {'object': Object()};
-        final normalizedValue = MethodChannelHelper.normalize(value);
+        final normalizedValue = normalize(value);
         when(channel.invokeMethod(
                 'setExtra', {'key': 'fixture-key', 'value': normalizedValue}))
             .thenAnswer((_) => Future.value());